test-tc-shuffle-cls_api.sh: Only fail for specific kernel
This is to avoid catching fixed issue by open issue checker.
The original bug was opened for old kernel with mlnx ofed but the fix
needs to be done in the kernel so the issue is marked as wont fix in
the issue tracker.
The comment was left for history but we dont want to catch it
as current kernels are fixed.
For the specific kernel it fails we now also fail the test.

Signed-off-by: Roi Dayan <roid@mellanox.com>