test-tc-max-rules.sh: Call reset_tc_nic after changing to legacy mode
With uplink representor, after changing from switchdev mode to legacy
mode, the PF netdev will be reloaded. The PF name could be the same
as the uplink representor, but it's a new netdev. We need to call
reset_tc_nic again to initialize it. Otherwise, we'll get the
following error:

    Error: TC offload is disabled on net device.

Change-Id: Ia0e6be46810b97b5f3e89a8e31591595a068fe20
Signed-off-by: Chris Mi <chrism@mellanox.com>