test-tc-vxlan-decap-in-skip-sw.sh: ignore by default
From RM:

Following a community discussion it was decided that this is a feature and not a
bug. In other words, skip_sw doesn't make sense to ingress (decap) rules in
installed on SW tunnel devices, b/c the packet can be forwarded by the kernel to
the tunnel device from anywhere (e.g not strictly according to routing as done
for the encap case). I suggest to close this case as "won't fix" or alike
status.

To make it clear, this should not be the case for vlan push/pop rules.

Change-Id: I1c839e827a3a4e203c2506242fe03970b8297fd1
Signed-off-by: Roi Dayan <roid@mellanox.com>