test-ethtool-basic.sh: Fix test not catching errors on second run
ethtool doesn't seem to fail changing hfunc but there is
are syndromes in the log. so catch it.
also reset back the hfunc to second run will catch the syndrome again.
also test 2 rss hkeys to make sure we change it and change back.
also reset back the equal value to 1.

Change-Id: I17e924a0919ceee6890a7cc67bf0df6141cfb7f3
Signed-off-by: Roi Dayan <roid@mellanox.com>