setup_configure.py: Add second-server argument
This argument indicates we config a second server to act
as a remote server which means it doesn't need full configuration.
For now only skip the end of the config, creating config file.
Later we can move VFs configuration to the end which is not needed
as well, need to verify functions depending on it like the bring interface
up don't break. i.e. it will bring PFs up but no VFs or reps which won't exists.

Change-Id: I6d2f303b33d6d919a5739afc65f32f3f2cdf259c
Signed-off-by: Roi Dayan <roid@mellanox.com>