test-eswitch-add/del-flows-during-reload.sh: Don't fail if reload fails
On some ofed versions we have custom cls_flower module which will take
refcount on mlx5_core and thus reload modules will fail.
We just want to load back unloaded modules so make sure next tests are ok
but we should not fail the current test.
The tests created with expectation not to crash.

Change-Id: Ifeb81e24e83e11a5514455c501c918f213a5348c
Signed-off-by: Roi Dayan <roid@mellanox.com>