ovs-dpdk-tests: Fix test-ovs-dpdk-e2e-del-flows
Due to new ovs code which uses lockless queues instead of
mutex between the insertion and deletion threads, this causes
the main thread to not lock the queue, and won't wait for the e2e-thread
to finish deleting before returning since the deletion is async.
Update the test accordingly by sleeping one second before querying the
merged flow to allow the completion of flow deletion.

Change-Id: Ic5b6c622efb6a668e86b09a121f1bb70bbf86045
Signed-off-by: Salem Sol <salems@nvidia.com>