common.sh: Disable sriov on port2 when test starts
Sometimes there is a bug when port2 is also configured but speific
tests should take care of testing 2 ports and not have random result
of how the system state is. so for tests to be consistent we make sure
port2 is not configured as sriov and tests that needs port2 will config.

Issue: 1824795
Change-Id: If3792d0f017ada085cb6ee7aa029280877528f86
Signed-off-by: Roi Dayan <roid@nvidia.com>